Added typescript and prisma
